minor text changes

+++ choose.cpp
@@ -1,4 +1,8 @@
 #include "permute.cpp"
-int choose(int n, int k){
+long int choose(long int n, long int k){
 	return permute(n,k)/permute(k,k);
 }
+
+
+
+
