+++ multichoose.cpp
@@ -1,4 +1,4 @@
 #include "choose.cpp"
-int multichoose(int n, int k){
+long int multichoose(int n, int k){
 	return choose(n+k-1, k);
 }
